feat: Add character sheet components and functionality
- Implemented AddCompetenceModal for adding new competences.
- Created CharacterSheet component to manage character data and tabs for stats, equipment, and description.
- Developed CompetencesPanel to display and manage character competences.
- Added DescriptionPanel for character details and custom fields.
- Built EquipPanel for managing weapons, armor, and items.
- Integrated ImportExportMenu for saving and loading character data.
- Added LevelUpPanel for handling character leveling up mechanics.
- Created StatsPanel to display and edit character statistics and attack modifiers.
