+++ components/CharacterSheet/CompetencesPanel.tsx
