+++ components/CharacterSheet/DescriptionPanel.tsx
