+++ components/CharacterSheet/EquipPanel.tsx
